fix(test/perf): cast slot_mapping fixture to int32 in test_store_lowrank
UCStoreLowrank wrapper asserts block_idxs / token_idxs are int32
(matches mojo_store_lowrank_bf16 wheel kernel ABI). The fixture
was generating int64 (default torch.randperm dtype on CPU), causing
all 14 perf shapes to FAIL with dtype assertion.

Unblocks ~14 rows of UCStoreLowrank perf data in subsequent R5
per-op surveys.

+++ mojo_opset/tests/perf/test_store_lowrank.py
@@ -25,8 +25,8 @@ shapes_key_lr = [
         (
             torch.zeros(size=shape0, dtype=torch.bfloat16),
             torch.randn(size=(slot_mapping.shape[0], *shape1), dtype=torch.bfloat16),
-            slot_mapping // 512,
-            slot_mapping % 512,
+            (slot_mapping // 512).to(torch.int32),
+            (slot_mapping % 512).to(torch.int32),
             slot_mapping.shape[0],
         )
         for shape0, shape1 in zip(shapes_label_cache, shapes_key_lr)
